refactor(erros):  remove switch erros from util to right place in the code :hammer:
